requirements.txt
faiss-cpu==1.7.4
pyserini>=0.23.0
transformers==4.35.2
datasets==2.19.1
ipykernel==6.27.1
ipywidgets==8.1.1
tqdm==4.66.1
matplotlib==3.8.2
psycopg2-binary==2.9.9 # to save gooaq answers to postgres
pylcs==0.1.1
Unidecode==1.3.8
huggingface-hub==0.24.5
openai==1.33.0
plotnine==0.13.6
bertopic==0.16.3
sentence-transformers==3.0.1 # for bertopic
nltk==3.9.1 # for bertopic stopwords
cuml-cu12==24.8.0 --extra-index-url=https://pypi.nvidia.com # for bertopic
tiktoken==0.7.0
accelerate==0.34.2 # for training
pytrec_eval==0.5 # for training
deepspeed==0.15.1  # for training

# git+https://github.com/attardi/wikiextractor.git@v3.0.7 # NOTE we add submodule instead
